Latest libnklabs fixes

+++ src/nkarch_atsam.c
@@ -128,7 +128,7 @@ int nk_mcuflash_erase(const void *info, uint32_t address, uint32_t byte_count)
 	return rtn;
 }
 
-int nk_mcuflash_write(const void *info, uint32_t address, const uint8_t *data, uint32_t byte_count)
+int nk_mcuflash_write(const void *info, uint32_t address, const uint8_t *data, size_t byte_count)
 {
 	int rtn = 0; // Assume success
 	(void)info;
@@ -155,7 +155,7 @@ int nk_mcuflash_write(const void *info, uint32_t address, const uint8_t *data, u
 	return rtn;
 }
 
-int nk_mcuflash_read(const void *info, uint32_t address, uint8_t *data, uint32_t byte_count)
+int nk_mcuflash_read(const void *info, uint32_t address, uint8_t *data, size_t byte_count)
 {
 	int rtn = 0; // Assume success
 	(void)info;
